- Renamed forms module in common to django - Added CustomCheckbox field which makes use of the bootstrap 4 custom checkbox
