+++ {{cookiecutter.project_slug}}/common/templates/common/custom_checkbox.html
@@ -1,0 +1,10 @@
+{% raw -%}
+{% load crispy_forms_field %}
+
+<div class="form-group">
+  <div class="custom-control custom-checkbox">
+    {% crispy_field field 'class' 'custom-control-input' %}
+    <label class="custom-control-label" for="{{ field.id_for_label }}">{{ field.label }}</label>
+  </div>
+</div>
+{%- endraw %}
